+++ src/PublicNavigation.js
@@ -1,16 +1,7 @@
 import React from 'react';
-import { PropTypes as Types } from 'prop-types';
 
-const PublicNavigation = ({ inverse }) => (
-  <header className={`navbar${inverse ? ' navbar-inverse' : ''}`} />
-);
+import PublicNavigationWithoutContent from './PublicNavigationWithoutContent';
 
-PublicNavigation.propTypes = {
-  inverse: Types.bool,
-};
-
-PublicNavigation.defaultProps = {
-  inverse: false,
-};
+const PublicNavigation = props => <PublicNavigationWithoutContent {...props} />;
 
 export default PublicNavigation;