+++ src/PublicNavigationWithoutContent/index.js
@@ -1,0 +1,3 @@
+import PublicNavigationWithoutContent from './PublicNavigationWithoutContent';
+
+export default PublicNavigationWithoutContent;