+++ CHANGELOG.md
@@ -1,2 +1,12 @@
+# v0.2.0
+## Adds logo
+
+![image](https://user-images.githubusercontent.com/5443561/35562098-41198892-05aa-11e8-8952-4806593298b3.png)
+
+Passing the logo link (https://transferwise.com by default):
+```js
+<PublicNavigation logoLink="https://transferwise.com/borderful" />
+```
+
 # v0.1.0
 ## Initial release